[Alura PH] - Criacao da classe Funcionario Autenticavel
