+++ src/alura/bytebank/controller/SistemaInterno.java
@@ -1,13 +1,14 @@
 package alura.bytebank.controller;
 
+import alura.bytebank.model.FuncionarioAutenticavel;
 import alura.bytebank.model.Gerente;
 
 public class SistemaInterno {
 
     private int senha = 2222;
 
-    public void autentica(Gerente g) {
-        boolean autenticou = g.autentica(this.senha);
+    public void autentica(FuncionarioAutenticavel fa) {
+        boolean autenticou = fa.autentica(this.senha);
         if (autenticou) {
             System.out.println("Pode entrar no sistema!");
         } else {
