+++ src/alura/bytebank/model/Administrador.java
@@ -1,0 +1,9 @@
+package alura.bytebank.model;
+
+public class Administrador extends FuncionarioAutenticavel {
+
+    @Override public double getBonificacao() {
+        return 50;
+    }
+
+}
