+++ src/alura/bytebank/model/Funcionario.java
@@ -12,27 +12,27 @@ public abstract class  Funcionario {
         return nome;
     }
 
-    public Funcionario setNome(final String nome) {
+    public void setNome(String nome) {
         this.nome = nome;
-        return this;
+
     }
 
     public String getCpf() {
         return cpf;
     }
 
-    public Funcionario setCpf(final String cpf) {
+    public void setCpf(String cpf) {
         this.cpf = cpf;
-        return this;
+
     }
 
     public double getSalario() {
         return salario;
     }
 
-    public Funcionario setSalario(final double salario) {
+    public void setSalario(double salario) {
         this.salario = salario;
-        return this;
+
     }
 
 }
