+++ src/alura/bytebank/model/Gerente.java
@@ -1,21 +1,6 @@
 package alura.bytebank.model;
 
-public class Gerente extends Funcionario {
-
-    private int senha;
-
-    public void setSenha(int senha) {
-        this.senha = senha;
-    }
-
-    public boolean autentica(int senha) {
-        if (this.senha == senha) {
-            return true;
-        } else {
-            return  false;
-        }
-    }
-
+public class Gerente extends FuncionarioAutenticavel {
 
     public double getBonificacao() {
         System.out.println("Chamando o metodo de bonificacao do gerente");
